final version

+++ rtr/reccomend/models.py
@@ -3,7 +3,7 @@ from django.db import models
 class ClothingItem(models.Model):
     item_id = models.IntegerField(blank=True,primary_key=True)
     event_type = models.CharField(max_length=50)
-    avg_rating = models.IntegerField(blank=True, null=True)
+    avg_rating = models.FloatField(blank=True, null=True)
     img_url = models.CharField(max_length=200)
 
 class User(models.Model):
@@ -18,7 +18,7 @@ class Rating(models.Model):
     colthing_bust = models.IntegerField(blank=True, null=True)
     colthing_size = models.CharField(max_length=50)
     body_shape = models.CharField(max_length=50)
-    rating_val = models.IntegerField(blank=True, null=True)
+    rating_val = models.FloatField(blank=True, null=True)
     rating_text = models.CharField(max_length=200)
 
 class Event(models.Model):
